refactored post menu into its own component
