+++ src/Pages/TripDetails.tsx
@@ -4,31 +4,23 @@ import moment from "moment";
 import { opencageAPIkey } from "../config/constants";
 import { useSelector, useDispatch } from "react-redux";
 import { useParams } from "react-router-dom";
-import { showMessageWithTimeout } from "../Store/appState/actions";
 import { fetchSpecificTrip } from "../Store/trips/actions";
-import { endTrip } from "../Store/users/actions";
 import { selectIdTrip } from "../Store/trips/selector";
 import { selectUser, selectUserTrip } from "../Store/users/selector";
 import { selectAppLoading } from "../Store/appState/selector";
 import { Trip, Post, DefaultMarker } from "../Types/tripTypes";
 import PostCard from "../Components/postCard/PostCard";
+import PostMenu from "../Components/postOptions/PostMenu";
 import GoogleMaps from "../Components/GoogleMaps";
-import NewPostModal from "../Components/NewPostModal";
 import Loading from "../Components/AppStateComponents/Loading";
 import {
   Flex,
   Box,
   Heading,
-  Button,
   TabList,
   Tab,
   Tabs,
   TabPanels,
-  Collapse,
-  FormControl,
-  Input,
-  InputGroup,
-  FormLabel,
 } from "@chakra-ui/core";
 import "../Style/GenStyle.css";
 
@@ -40,8 +32,6 @@ export default function TripDetails() {
   const userTrip = useSelector(selectUserTrip(id));
   const tripFromTripReducer = useSelector(selectIdTrip(id));
   const oneTrip: Trip = userTrip ? userTrip : tripFromTripReducer;
-  const [toggle_endDate, set_toggle_endDate] = useState(false);
-  const [endDate, set_endDate] = useState("");
   const [postIndex, set_postIndex] = useState(0);
   const [moveToMarker, set_moveToMarker] = useState<DefaultMarker>({
     lat: 0,
@@ -62,80 +52,10 @@ export default function TripDetails() {
     });
   }
 
-  function tripControlMenu() {
-    return (
-      <>
-        <Flex width="30vw" margin="auto" justify="space-around" mb="2rem">
-          <NewPostModal />
-          {!oneTrip.endDate ? (
-            <Button
-              className="navbtn"
-              variantColor="customBtn"
-              onClick={(e) => set_toggle_endDate(!toggle_endDate)}
-            >
-              End Trip
-            </Button>
-          ) : null}
-        </Flex>
-        <Collapse mt={4} isOpen={toggle_endDate} m="auto" w="15vw" mb={6}>
-          <FormControl>
-            <InputGroup maxH="10vh">
-              <Flex
-                w="5vw"
-                d="column"
-                style={{ height: "10vh" }}
-                m="auto"
-                mt={2}
-              >
-                <FormLabel htmlFor="text">Please confirm the date</FormLabel>
-                <Input
-                  type="date"
-                  variant="flushed"
-                  onChange={(e: any) => set_endDate(e.target.value)}
-                />
-              </Flex>
-            </InputGroup>
-          </FormControl>
-          <Flex>
-            <Button
-              minW="10vw"
-              maxW="10vw"
-              variantColor="customBtn"
-              className="navbtn"
-              m="auto"
-              onClick={(e) => endtrip(e)}
-            >
-              Confirm
-            </Button>
-          </Flex>
-        </Collapse>
-      </>
-    );
-  }
-
   function tabOnClickHandler(tabIndex: number, post: Post) {
     const { latitude, longitude } = post;
     reverseGeoCode(latitude, longitude);
     set_postIndex(tabIndex);
-  }
-
-  function endtrip(e: any) {
-    // e.preventDefault();
-    if (!endDate) {
-      const message = "Please pick a date to end your trip";
-      dispatch(showMessageWithTimeout("error", true, message, 3000));
-    } else {
-      const momentStartTripDate = moment(oneTrip.startDate);
-      const momentEndTripDate = moment(endDate);
-      if (momentStartTripDate > momentEndTripDate) {
-        const message = "Trip can not end before the date it started!";
-        dispatch(showMessageWithTimeout("error", true, message, 3000));
-      } else {
-        const endingTrip = { ...oneTrip, endDate };
-        dispatch(endTrip(endingTrip));
-        set_toggle_endDate(!toggle_endDate);
-      }
-    }
   }
 
   useEffect(() => {
@@ -148,13 +68,14 @@ export default function TripDetails() {
   const isTripOver = () => {
     if (oneTrip.endDate) {
       const momentCurrentDate = moment(new Date());
-      const momentEndTripDate = moment(endDate);
+      const momentEndTripDate = moment(oneTrip.endDate);
       if (momentCurrentDate > momentEndTripDate) return true;
     }
     return false;
   };
   const validation = oneTrip && user.token && !isTripOver() && isUser;
-  const menu = validation ? tripControlMenu() : null;
+  console.log("validation", isUser);
+  const menu = validation ? <PostMenu oneTrip={oneTrip} /> : null;
 
   if (loading) {
     return <Loading />;
